+++ src/types/index.ts
@@ -5,6 +5,7 @@
 export * from "./auth.types";
 export * from "./sleep.types";
 export * from "./fasting.types";
+export * from "./settings.types";
 export * from "./caffeine.types";
 export * from "./todo.types";
 export * from "./notification.types";